Extract some common types from `spi::master` into the root `spi` module
